requirements.txt
pydantic
pytest
faiss-cpu
numpy
pypdf
openai
python-dotenv